Remove some `expect` Statements
In these cases, the expect statements were basically assertions.
Nonetheless, this lib should never panic and therefore it makes sense
to remove these expect statements.

An alternative would be to wrap the top-level functions in try blocks and
map to an Error whenever we encounter a panic.
However, I want to try and stick with the statically checked result,
ensuring all unexpected errors are handled.