Switch to CRC32-based Hash
Rather than extracting and hashing the contents of a GDTF file, this
uses the CRC32 that is contained for every file in a zip. This improves
performance, such that `cargo test` goes from about 20 s to 14 s,
which is a speed improvement of around 30 %.

Collision probability with a 32-bit hash should still be on the order
of 1e-6, which seems sufficient for a low risk application like this.